api/requirements.txt
aniso8601==8.1.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
passlib==1.7.4
pycparser==2.20
PyJWT==2.0.0
PySocks==1.7.1
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.20
tweepy==3.9.0
urllib3==1.26.2
Werkzeug==1.0.1
